Moved batchnorm related TPPs into separate headers
