+++ samples/deeplearning/fused_bn_gn_tpp/layer_example_f32.c
@@ -16,807 +16,13 @@
 #include <math.h>
 #include <omp.h>
 
-#define BITS_PER_CHAR (8)
-
-#define NUM_HW_BLOCKS (16)
+/* include c-based dnn library */
+#include "../common/dnn_common.h"
+#include "bn_tpp_common.h"
+#include "bn_tpp_fwd_custom_f32.h"
+#include "bn_tpp_bwd_custom_f32.h"
 
 #define COMPUTE_FP64_REFERENCE
-
-/* include c-based dnn library */
-#ifdef CNN_HEADER
-  #include "dnn_common.h"
-#else /* LIBXSMM sample */
-  #include "../common/dnn_common.h"
-#endif
-
-typedef enum my_normalization_fuse {
-  MY_NORMALIZE_FUSE_NONE = 0,
-  MY_NORMALIZE_FUSE_RELU = 1,
-  MY_NORMALIZE_FUSE_ELTWISE = 2,
-  MY_NORMALIZE_FUSE_ELTWISE_RELU = 3,
-  MY_NORMALIZE_FUSE_RELU_WITH_MASK = 4,
-  MY_NORMALIZE_FUSE_ELTWISE_RELU_WITH_MASK = 5
-} my_normalization_fuse;
-
-typedef struct my_bn_fwd_config {
-  libxsmm_blasint  N;
-  libxsmm_blasint  C;
-  libxsmm_blasint  H;
-  libxsmm_blasint  W;
-  libxsmm_blasint  bc;
-  libxsmm_blasint  CP;
-  libxsmm_blasint  num_HW_blocks;
-  libxsmm_blasint  threads;
-  size_t           scratch_size;
-
-  libxsmm_barrier* barrier;
-
-  libxsmm_matrix_eqn_function  func10;
-  libxsmm_meltwfunction_unary  reduce_HW_kernel;
-  libxsmm_meltwfunction_unary  all_zero_kernel;
-  libxsmm_meltwfunction_binary add_kernel; // FIXME: rename since this is a helper not the true inp_add */
-  libxsmm_meltwfunction_unary  copy_kernel;// FIXME: rename since this is a helper not the true inp_add */
-  libxsmm_meltwfunction_unary  relu_kernel;
-  libxsmm_meltwfunction_binary ewise_add_kernel;
-  my_normalization_fuse        fuse_type;
-} my_bn_fwd_config;
-
-typedef struct my_bn_bwd_config {
-  libxsmm_blasint  N;
-  libxsmm_blasint  C;
-  libxsmm_blasint  H;
-  libxsmm_blasint  W;
-  libxsmm_blasint  bc;
-  libxsmm_blasint  CP;
-  libxsmm_blasint  num_HW_blocks;
-  libxsmm_blasint  threads;
-  size_t           scratch_size;
-
-  libxsmm_barrier* barrier;
-
-  libxsmm_matrix_eqn_function  dgamma_func;
-  libxsmm_matrix_eqn_function  dbeta_func;
-  libxsmm_matrix_eqn_function  din_func;
-  libxsmm_meltwfunction_unary  all_zero_kernel;
-  libxsmm_meltwfunction_binary add_kernel; // FIXME: rename since this is a helper not the true inp_add */
-  libxsmm_meltwfunction_unary  copy_kernel;// FIXME: rename since this is a helper not the true inp_add */
-  libxsmm_meltwfunction_unary  relu_kernel;
-  libxsmm_meltwfunction_binary ewise_add_kernel;
-  libxsmm_meltwfunction_unary  ewise_copy_kernel;
-  my_normalization_fuse        fuse_type;
-} my_bn_bwd_config;
-
-my_bn_fwd_config setup_my_bn_fwd(libxsmm_blasint N, libxsmm_blasint C, libxsmm_blasint H, libxsmm_blasint W, libxsmm_blasint bc,
-                                 libxsmm_blasint threads, my_normalization_fuse fuse_type ) {
-  my_bn_fwd_config res;
-
-  size_t sum_N_offset, sumsq_N_offset;
-
-  libxsmm_blasint ldo = bc;
-  libxsmm_blasint ld  = bc;
-  libxsmm_blasint tmp_ld, tmp_ld2;
-  libxsmm_blasint my_eqn10;
-
-  libxsmm_meltw_unary_flags jit_reduce_flags = LIBXSMM_MELTW_FLAG_UNARY_NONE;
-  libxsmm_meltw_unary_type  unary_type;
-
-  libxsmm_datatype  in_dt  = LIBXSMM_DATATYPE_F32;
-  libxsmm_datatype  out_dt = LIBXSMM_DATATYPE_F32;
-
-  /* setting up some handle values */
-  res.N  = N;
-  res.C  = C;
-  res.H  = H;
-  res.W  = W;
-  res.bc = bc;
-  res.CP = res.C / res.bc;
-  res.num_HW_blocks = NUM_HW_BLOCKS; /* hardcoded for now */
-  res.threads       = threads;
-  res.fuse_type     = fuse_type;
-
-  /* when masking is on, bc must be divisible by 8 for compressing mask into char array (otherwise strides are wrong for relumask */
-  if ( (res.fuse_type == 4 || res.fuse_type == 5) && (res.bc % BITS_PER_CHAR != 0)) {
-    fprintf( stderr, "bc = %d is not divisible by BITS_PER_CHAR = %d. Bailing...!\n", res.bc, BITS_PER_CHAR);
-    exit(-1);
-  }
-
-  /* setting up the barrier */
-  res.barrier = libxsmm_barrier_create(threads, 1);
-
-  /* TPP creation */
-
-  /* Eltwise TPPs  */
-  res.all_zero_kernel = libxsmm_dispatch_meltw_unary(res.bc, 1, NULL, &ldo, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_MELTW_FLAG_UNARY_NONE, LIBXSMM_MELTW_TYPE_UNARY_XOR);
-  if ( res.all_zero_kernel == NULL) {
-    fprintf( stderr, "JIT for TPP fwd all_zero_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  res.copy_kernel = libxsmm_dispatch_meltw_unary(res.bc, 1, &ldo, &ldo, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_MELTW_FLAG_UNARY_NONE, LIBXSMM_MELTW_TYPE_UNARY_IDENTITY);
-  if ( res.copy_kernel == NULL) {
-    fprintf( stderr, "JIT for TPP fwd copy_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  res.add_kernel = libxsmm_dispatch_meltw_binary(res.bc, 1, &ldo, &ldo, &ldo, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_MELTW_FLAG_BINARY_NONE, LIBXSMM_MELTW_TYPE_BINARY_ADD);
-  if ( res.add_kernel == NULL) {
-    fprintf( stderr, "JIT for TPP fwd add_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  res.relu_kernel = libxsmm_dispatch_meltw_unary(res.bc, res.H*res.W / res.num_HW_blocks, &ldo, &ldo, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32,
-                                                 /*LIBXSMM_MELTW_FLAG_UNARY_NONE*/ LIBXSMM_MELTW_FLAG_UNARY_BITMASK_2BYTEMULT, LIBXSMM_MELTW_TYPE_UNARY_RELU);
-  if ( res.relu_kernel == NULL ) {
-    fprintf( stderr, "JIT for TPP fwd_relu_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  res.ewise_add_kernel = libxsmm_dispatch_meltw_binary(res.bc, res.H*res.W / res.num_HW_blocks, &ldo, &ldo, &ldo, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32,
-                                                       LIBXSMM_MELTW_FLAG_BINARY_NONE, LIBXSMM_MELTW_TYPE_BINARY_ADD);
-  if ( res.ewise_add_kernel == NULL) {
-    fprintf( stderr, "JIT for TPP fwd ewise_add_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  /* TPPs for reducing X and X2 in HW*/
-  tmp_ld = bc;
-
-  unary_type = LIBXSMM_MELTW_TYPE_UNARY_REDUCE_X_X2_OP_ADD;
-  jit_reduce_flags = LIBXSMM_MELTW_FLAG_UNARY_REDUCE_COLS;
-  res.reduce_HW_kernel = libxsmm_dispatch_meltw_unary(res.bc, res.H*res.W/res.num_HW_blocks, &ld, &tmp_ld, in_dt, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, jit_reduce_flags, unary_type);
-  if ( res.reduce_HW_kernel == NULL) {
-    fprintf( stderr, "JIT for TPP fwd reduce_HW_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  /* TPP for scaling */
-  ld = bc;
-  tmp_ld = 1;
-  tmp_ld2 = 1;
-
-  my_eqn10 = libxsmm_matrix_eqn_create();                                                            /* y = (s*x + b)*gamma + beta */
-  libxsmm_matrix_eqn_push_back_ternary_op( my_eqn10, LIBXSMM_MELTW_TYPE_TERNARY_MULADD, LIBXSMM_MELTW_FLAG_TERNARY_BCAST_COL_IN_1 | LIBXSMM_MELTW_FLAG_TERNARY_BCAST_COL_IN_2 | LIBXSMM_MELTW_FLAG_TERNARY_REUSE_IN_2_AS_OUT, LIBXSMM_DATATYPE_F32 );
-  libxsmm_matrix_eqn_push_back_ternary_op( my_eqn10, LIBXSMM_MELTW_TYPE_TERNARY_MULADD, LIBXSMM_MELTW_FLAG_TERNARY_BCAST_COL_IN_1 | LIBXSMM_MELTW_FLAG_TERNARY_BCAST_COL_IN_2 | LIBXSMM_MELTW_FLAG_TERNARY_REUSE_IN_2_AS_OUT, LIBXSMM_DATATYPE_F32 );
-  libxsmm_matrix_eqn_push_back_arg( my_eqn10, res.bc, res.H*res.W /res.num_HW_blocks, ld, 0, 0, in_dt );   /* x = [HW, bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn10, res.bc, 1, tmp_ld, 1, 0, LIBXSMM_DATATYPE_F32 );       /* s = [bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn10, res.bc, 1, tmp_ld, 2, 0, LIBXSMM_DATATYPE_F32 );       /* b = [bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn10, res.bc, 1, tmp_ld2, 3, 0, in_dt );                     /* gamma = [bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn10, res.bc, 1, tmp_ld2, 4, 0, in_dt );                     /* beta = [bc] */
-  /* libxsmm_matrix_eqn_tree_print( my_eqn10 ); */
-  /* libxsmm_matrix_eqn_rpn_print( my_eqn10 ); */
-
-  res.func10 = libxsmm_dispatch_matrix_eqn( res.bc, res.H*res.W / res.num_HW_blocks, &ld, out_dt, my_eqn10 );   /* y = [HW, bc] */
-  if ( res.func10 == NULL) {
-    fprintf( stderr, "JIT for TPP fwd func10 (eqn10) failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  /* init scratch */
-  sum_N_offset   = LIBXSMM_UP2(res.CP * 2 * res.bc, 64);
-  sumsq_N_offset = LIBXSMM_UP2(sum_N_offset + res.CP * res.N * res.bc, 64);
-  res.scratch_size =  sizeof(float) * ( sumsq_N_offset /*sum_X_X2 + sumsq_N */ + LIBXSMM_UP2((size_t)res.CP * (size_t)res.N * (size_t)res.bc, 64) /* sumsq_N */ );
-
-  return res;
-}
-
-void my_bn_fwd_exec( my_bn_fwd_config cfg, const float *pinp, const float *pinp_add, const float *pgamma, const float *pbeta, float *mean, float *var, float *pout, unsigned char *prelumask, float eps, int start_tid, int my_tid, void *scratch ) {
-
-  const libxsmm_blasint N  = cfg.N;
-  const libxsmm_blasint CP = cfg.CP;
-  const libxsmm_blasint HW = cfg.H * cfg.W;
-  const libxsmm_blasint bc = cfg.bc;
-  const libxsmm_blasint num_HW_blocks = cfg.num_HW_blocks;
-
-  /* computing first logical thread */
-  const libxsmm_blasint ltid = my_tid - start_tid;
-
-  /* number of tasks that could be run in parallel for 1d blocking */
-  // Question: each thread should take a number of full (of length CP chunks) or can we really do a partial split here?
-  const libxsmm_blasint work_dN = CP * N;
-  /* compute chunk size */
-  const libxsmm_blasint chunksize_dN = (work_dN % cfg.threads == 0) ?
-    (work_dN / cfg.threads) : ((work_dN / cfg.threads) + 1);
-  /* compute thr_begin and thr_end */
-  const libxsmm_blasint thr_begin_dN = (ltid * chunksize_dN < work_dN) ? (ltid * chunksize_dN) : work_dN;
-  const libxsmm_blasint thr_end_dN = ((ltid + 1) * chunksize_dN < work_dN) ? ((ltid + 1) * chunksize_dN) : work_dN;
-
-  /* number of tasks that could be run in parallel for 1d blocking */
-  // Question: each thread should take a number of full (of length CP chunks) or can we really do a partial split here?
-  const libxsmm_blasint work_C = CP;
-  /* compute chunk size */
-  const libxsmm_blasint chunksize_C = (work_C % cfg.threads == 0) ?
-    (work_C / cfg.threads) : ((work_C / cfg.threads) + 1);
-  /* compute thr_begin and thr_end */
-  const libxsmm_blasint thr_begin_C = (ltid * chunksize_C < work_C) ? (ltid * chunksize_C) : work_C;
-  const libxsmm_blasint thr_end_C = ((ltid + 1) * chunksize_C < work_C) ? ((ltid + 1) * chunksize_C) : work_C;
-
-  /* lazy barrier init */
-  libxsmm_barrier_init(cfg.barrier, ltid);
-
-  LIBXSMM_VLA_DECL(4, const float,         inp,      pinp, CP, HW, bc);            /* [N, CP, HW, bc] */
-  LIBXSMM_VLA_DECL(4,       float,         out,      pout, CP, HW, bc);            /* [N, CP, HW, bc] */
-  LIBXSMM_VLA_DECL(2, const float,         gamma,    pgamma, bc);                  /* [CP, bc] */
-  LIBXSMM_VLA_DECL(2, const float,         beta,     pbeta, bc);                   /* [CP, bc] */
-
-  LIBXSMM_VLA_DECL(4, const float,         inp_add,  pinp_add, CP, HW, bc);        /* [N, CP, HW, bc] */
-
-  float alpha = 0.0f;
-  LIBXSMM_VLA_DECL(4,       unsigned char, relumask, prelumask, CP, HW, bc/BITS_PER_CHAR);    /* [N, CP, HW, bc/BITS_PER_CHAR] */
-
-  const float scale = 1.0f /((float)N * HW);
-
-  LIBXSMM_VLA_DECL(3, float, sum_X_X2, ((float*)scratch), CP, bc);  /* [2, CP, bc] */
-  LIBXSMM_ASSUME_ALIGNED(sum_X_X2_, 64);
-  const libxsmm_blasint sum_N_offset = (LIBXSMM_UP2((uintptr_t)(((float*)scratch) + CP * 2 * bc), 64) - ((uintptr_t)(scratch))) / sizeof(float);
-  LIBXSMM_VLA_DECL(3, float, sum_N, ((float*)scratch) + sum_N_offset, N, bc);  /* [CP, N, bc] */
-  LIBXSMM_ASSUME_ALIGNED(sum_N_, 64);
-  const libxsmm_blasint sumsq_N_offset = (LIBXSMM_UP2((uintptr_t)(((float*)scratch) + sum_N_offset + CP * N * bc), 64) - ((uintptr_t)(scratch))) / sizeof(float);
-  LIBXSMM_VLA_DECL(3, float, sumsq_N, ((float*)scratch) + sumsq_N_offset, N, bc);  /* [CP, N, bc] */
-  LIBXSMM_ASSUME_ALIGNED(sumsq_N_, 64);
-
-  { /* stupid block to keep indentation */
-    LIBXSMM_ALIGNED(float s[bc], 64);
-    LIBXSMM_ALIGNED(float b[bc], 64);
-    int n, cp;
-
-    int cpxnt;
-    for ( cpxnt = thr_begin_dN; cpxnt < thr_end_dN; ++cpxnt ) {
-      { /* stupid block to keep indentation */
-        n  = cpxnt%N;
-        cp = cpxnt/N;
-
-        int hwb;
-
-        float *sum_ncp_ptr   = &LIBXSMM_VLA_ACCESS(3, sum_N, cp, n, 0, N, bc);
-        float *sumsq_ncp_ptr = &LIBXSMM_VLA_ACCESS(3, sumsq_N, cp, n, 0, N, bc);
-
-        libxsmm_meltw_unary_param all_zero_param;
-        all_zero_param.out.primary = sum_ncp_ptr;
-        cfg.all_zero_kernel(&all_zero_param);
-        all_zero_param.out.primary = sumsq_ncp_ptr;
-        cfg.all_zero_kernel(&all_zero_param);
-
-        /* #pragma omp simd  */
-        /* for (int cb = 0; cb < bc; cb++) {  */
-        /*   sum_ncp_ptr[cb] = 0.0f;    */
-        /*   sumsq_ncp_ptr[cb] = 0.0f;  */
-        /* } */
-
-        libxsmm_meltw_binary_param add_param;
-
-        libxsmm_meltw_unary_param reduce_HW_params;       /*Private params and tmp array */
-        LIBXSMM_ALIGNED(float lcl_sum_X_X2[2*bc], 64);
-        reduce_HW_params.out.primary   = lcl_sum_X_X2;                                                         /* [2*bc]  */
-        for(hwb=0; hwb < num_HW_blocks; hwb++){
-
-          reduce_HW_params.in.primary = (void*)&LIBXSMM_VLA_ACCESS(4, inp, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);
-          cfg.reduce_HW_kernel(&reduce_HW_params);                                                       /* [HW, bc] -----> [2 * bc] */
-
-          add_param.in0.primary = sum_ncp_ptr;
-          add_param.in1.primary = lcl_sum_X_X2;
-          add_param.out.primary = sum_ncp_ptr;
-          cfg.add_kernel(&add_param);
-
-          add_param.in0.primary = sumsq_ncp_ptr;
-          add_param.in1.primary = &lcl_sum_X_X2[bc];
-          add_param.out.primary = sumsq_ncp_ptr;
-          cfg.add_kernel(&add_param);
-
-          /* #pragma omp simd */
-          /* for (int cb = 0; cb < bc; cb++) {  */
-          /*   sum_ncp_ptr[cb] += lcl_sum_X_X2[cb];  */
-          /*   sumsq_ncp_ptr[cb] += lcl_sum_X_X2[bc + cb];  */
-          /* }  */
-        }
-      }
-    }
-
-    libxsmm_barrier_wait(cfg.barrier, ltid);
-
-    for ( cp = thr_begin_C; cp < thr_end_C; ++cp ) {
-
-      libxsmm_meltw_unary_param all_zero_param;
-      all_zero_param.out.primary = &LIBXSMM_VLA_ACCESS(3, sum_X_X2, 0, cp, 0, CP, bc);
-      cfg.all_zero_kernel(&all_zero_param);
-      all_zero_param.out.primary = &LIBXSMM_VLA_ACCESS(3, sum_X_X2, 1, cp, 0, CP, bc);
-      cfg.all_zero_kernel(&all_zero_param);
-
-      /* #pragma omp simd */
-      /* for (int cb = 0; cb < bc; cb++) {  */
-      /*   sum_X_X2[cp*bc + cb] = 0.0f;   */
-      /*   sum_X_X2[CP*bc + (cp*bc + cb)] = 0.0f;  */
-      /* } */
-
-      libxsmm_meltw_binary_param add_param;
-      int cb, ni;
-      for(ni = 0; ni < N; ni++){
-
-        add_param.in0.primary = &LIBXSMM_VLA_ACCESS(3, sum_X_X2, 0, cp, 0, CP, bc);
-        add_param.in1.primary = &LIBXSMM_VLA_ACCESS(3, sum_N, cp, ni, 0, N, bc);
-        add_param.out.primary = &LIBXSMM_VLA_ACCESS(3, sum_X_X2, 0, cp, 0, CP, bc);
-        cfg.add_kernel(&add_param);
-
-        add_param.in0.primary = &LIBXSMM_VLA_ACCESS(3, sum_X_X2, 1, cp, 0, CP, bc);
-        add_param.in1.primary = &LIBXSMM_VLA_ACCESS(3, sumsq_N, cp, ni, 0, N, bc);
-        add_param.out.primary = &LIBXSMM_VLA_ACCESS(3, sum_X_X2, 1, cp, 0, CP, bc);
-        cfg.add_kernel(&add_param);
-
-        /* #pragma omp simd */
-        /* for (int cb = 0; cb < bc; cb++) { */
-        /*   sum_X_X2[cp*bc + cb] += sum_N[cp*N*bc + n*bc + cb]; */
-        /*   sum_X_X2[CP*bc + (cp*bc + cb)] += sumsq_N[cp*N*bc + n*bc + cb]; */
-        /* } */
-      }
-
-      for(cb = 0; cb < bc; cb++){
-        mean[cp*bc + cb] = (LIBXSMM_VLA_ACCESS(3, sum_X_X2, 0, cp, cb, CP, bc)) * scale;                 /* E[X] */
-        var[cp*bc + cb] = ((LIBXSMM_VLA_ACCESS(3, sum_X_X2, 1, cp, cb, CP, bc)) * scale) - (mean[cp*bc + cb]*mean[cp*bc + cb]);
-      }
-    }
-
-    libxsmm_barrier_wait(cfg.barrier, ltid);
-
-    for ( cpxnt = thr_begin_dN; cpxnt < thr_end_dN; ++cpxnt ) {
-      { /* stupid block to keep indentation */
-        n  = cpxnt%N;
-        cp = cpxnt/N;
-
-        libxsmm_matrix_arg arg_array[5];                                                         /* private eqn args and params*/
-        libxsmm_matrix_eqn_param eqn_param;
-        int hwb, cb;
-
-        for(cb = 0; cb < bc; cb++){
-          s[cb] = 1.0f / ((float)sqrt(var[cp*bc + cb] + eps));                                 /* s = 1/sqrt(var(X) + eps)     [bc] */
-          b[cb] = -1 * mean[cp*bc + cb] * s[cb];                                               /* b = -E[X]/sqrt(var(X) + eps) [bc] */
-        }
-        arg_array[1].primary = s;                                                              /* [bc] */
-        arg_array[2].primary = b;                                                              /* [bc] */
-        arg_array[3].primary = (void*)&LIBXSMM_VLA_ACCESS(2, gamma, cp, 0, bc);                       /* [bc] */
-        arg_array[4].primary = (void*)&LIBXSMM_VLA_ACCESS(2, beta,  cp, 0, bc);                       /* [bc] */
-
-        for(hwb=0; hwb < num_HW_blocks; hwb++){
-
-          arg_array[0].primary = (void*)&LIBXSMM_VLA_ACCESS(4, inp, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);           /* [HW, bc] */
-          eqn_param.inputs = arg_array;
-          eqn_param.output.primary = &LIBXSMM_VLA_ACCESS(4, out, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);       /* [HW,bc] */
-          cfg.func10(&eqn_param);                                                                    /* Normalization equation -> y = ((s*x + b)*gamma + beta) */
-
-          /* Eltwise add */
-          if (cfg.fuse_type == MY_NORMALIZE_FUSE_ELTWISE || cfg.fuse_type == MY_NORMALIZE_FUSE_ELTWISE_RELU_WITH_MASK) {
-            libxsmm_meltw_binary_param add_param;
-            add_param.in0.primary = (void*)&LIBXSMM_VLA_ACCESS(4, out,     n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);
-            add_param.in1.primary = (void*)&LIBXSMM_VLA_ACCESS(4, inp_add, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);
-            add_param.out.primary = (void*)&LIBXSMM_VLA_ACCESS(4, out,     n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);
-            cfg.ewise_add_kernel(&add_param);
-          }
-
-          /* ReLU */
-          if (cfg.fuse_type == MY_NORMALIZE_FUSE_RELU || cfg.fuse_type == MY_NORMALIZE_FUSE_RELU_WITH_MASK || cfg.fuse_type == MY_NORMALIZE_FUSE_ELTWISE_RELU_WITH_MASK) {
-            libxsmm_meltw_unary_param all_relu_param;
-
-            all_relu_param.op.primary   = (void*)(&alpha);
-            all_relu_param.in.primary   = &LIBXSMM_VLA_ACCESS(4, out, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);      /* [HW,bc] */
-            all_relu_param.out.primary  = &LIBXSMM_VLA_ACCESS(4, out, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);      /* [HW,bc] */
-            all_relu_param.out.secondary = ((cfg.fuse_type == MY_NORMALIZE_FUSE_RELU_WITH_MASK || cfg.fuse_type == MY_NORMALIZE_FUSE_ELTWISE_RELU_WITH_MASK) ?
-                                              (void*)&LIBXSMM_VLA_ACCESS(4, relumask, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc/BITS_PER_CHAR) : NULL );
-            cfg.relu_kernel(&all_relu_param);
-          } /* ReLU */
-        }
-      }
-    }
-  }
-
-  libxsmm_barrier_wait(cfg.barrier, ltid);
-
-}
-
-my_bn_bwd_config setup_my_bn_bwd(libxsmm_blasint N, libxsmm_blasint C, libxsmm_blasint H, libxsmm_blasint W, libxsmm_blasint bc,
-                                 libxsmm_blasint threads, my_normalization_fuse fuse_type ) {
-  my_bn_bwd_config res;
-
-  size_t dbeta_N_offset;
-
-  libxsmm_blasint ldo = bc;
-  libxsmm_blasint ld  = bc;
-  libxsmm_blasint tmp_ld2;
-  libxsmm_blasint my_eqn11, my_eqn12, my_eqn16;
-
-  libxsmm_datatype  in_dt  = LIBXSMM_DATATYPE_F32;
-  libxsmm_datatype  out_dt = LIBXSMM_DATATYPE_F32;
-
-  /* setting up some handle values */
-  res.N  = N;
-  res.C  = C;
-  res.H  = H;
-  res.W  = W;
-  res.bc = bc;
-  res.CP = res.C / res.bc;
-  res.num_HW_blocks = NUM_HW_BLOCKS; /* hardcoded for now */
-  res.threads       = threads;
-  res.fuse_type     = fuse_type;
-
-  /* when masking is on, bc must be divisible by 8 for compressing mask into char array (otherwise strides are wrong for relumask */
-  if ( (res.fuse_type == 4 || res.fuse_type == 5) && (res.bc % BITS_PER_CHAR != 0)) {
-    fprintf( stderr, "bc = %d is not divisible by BITS_PER_CHAR = %d. Bailing...!\n", res.bc, BITS_PER_CHAR);
-    exit(-1);
-  }
-
-  /* setting up the barrier */
-  res.barrier = libxsmm_barrier_create(threads, 1);
-
-  /* TPP creation */
-
-  /* Eltwise TPPs  */
-  res.all_zero_kernel = libxsmm_dispatch_meltw_unary(res.bc, 1, NULL, &ldo, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32,
-                                                      LIBXSMM_MELTW_FLAG_UNARY_NONE, LIBXSMM_MELTW_TYPE_UNARY_XOR);
-  if ( res.all_zero_kernel == NULL) {
-    fprintf( stderr, "JIT for TPP bwd all_zero_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  res.copy_kernel = libxsmm_dispatch_meltw_unary(res.bc, 1, &ldo, &ldo, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32,
-                                                  LIBXSMM_MELTW_FLAG_UNARY_NONE, LIBXSMM_MELTW_TYPE_UNARY_IDENTITY);
-  if ( res.copy_kernel == NULL) {
-    fprintf( stderr, "JIT for TPP bwd copy_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  res.add_kernel = libxsmm_dispatch_meltw_binary(res.bc, 1, &ldo, &ldo, &ldo, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32,
-                                                  LIBXSMM_MELTW_FLAG_BINARY_NONE, LIBXSMM_MELTW_TYPE_BINARY_ADD);
-  if ( res.add_kernel == NULL) {
-    fprintf( stderr, "JIT for TPP bwd add_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  res.relu_kernel = libxsmm_dispatch_meltw_unary(res.bc, res.H*res.W / res.num_HW_blocks, &ldo, &ldo, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32,
-                                                 /*LIBXSMM_MELTW_FLAG_UNARY_NONE unsupported */ LIBXSMM_MELTW_FLAG_UNARY_BITMASK_2BYTEMULT, LIBXSMM_MELTW_TYPE_UNARY_RELU_INV);
-  if ( res.relu_kernel == NULL ) {
-    fprintf( stderr, "JIT for TPP bwd relu_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  res.ewise_add_kernel = libxsmm_dispatch_meltw_binary(res.bc, res.H*res.W / res.num_HW_blocks, &ldo, &ldo, &ldo, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32,
-                                                       LIBXSMM_MELTW_FLAG_BINARY_NONE, LIBXSMM_MELTW_TYPE_BINARY_ADD);
-  if ( res.ewise_add_kernel == NULL) {
-    fprintf( stderr, "JIT for TPP bwd ewise_add_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  res.ewise_copy_kernel = libxsmm_dispatch_meltw_unary(/*1, 1 */res.bc, res.H*res.W / res.num_HW_blocks, &ldo, &ldo, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32, LIBXSMM_DATATYPE_F32,
-                                                       LIBXSMM_MELTW_FLAG_UNARY_NONE, LIBXSMM_MELTW_TYPE_UNARY_IDENTITY);
-  if ( res.ewise_copy_kernel == NULL) {
-    fprintf( stderr, "JIT for TPP bwd ewise_copy_kernel failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  /* TPP equations for dgamma, dbeta and din */
-
-  ld = bc;
-  tmp_ld2 = 1;
-
-  /* dgamma function  */
-  my_eqn11 = libxsmm_matrix_eqn_create();                                                       /* dgamma = ((inp *a + b) * dout) + dgamma */
-  libxsmm_matrix_eqn_push_back_binary_op(my_eqn11, LIBXSMM_MELTW_TYPE_BINARY_ADD, LIBXSMM_MELTW_FLAG_BINARY_NONE, LIBXSMM_DATATYPE_F32);                   /* dgamma = ((inp *a + b) * dout) + dgamma */
-  libxsmm_matrix_eqn_push_back_unary_op(my_eqn11, LIBXSMM_MELTW_TYPE_UNARY_REDUCE_X_OP_ADD, LIBXSMM_MELTW_FLAG_UNARY_REDUCE_COLS, LIBXSMM_DATATYPE_F32);   /* [HW, bc] -> [bc] */
-  libxsmm_matrix_eqn_push_back_binary_op(my_eqn11, LIBXSMM_MELTW_TYPE_BINARY_MUL, LIBXSMM_MELTW_FLAG_BINARY_NONE, LIBXSMM_DATATYPE_F32);                   /* ((inp *a + b) * dout) */
-  libxsmm_matrix_eqn_push_back_ternary_op( my_eqn11, LIBXSMM_MELTW_TYPE_TERNARY_MULADD, LIBXSMM_MELTW_FLAG_TERNARY_BCAST_COL_IN_1 | LIBXSMM_MELTW_FLAG_TERNARY_BCAST_COL_IN_2 | LIBXSMM_MELTW_FLAG_TERNARY_REUSE_IN_2_AS_OUT, LIBXSMM_DATATYPE_F32 );
-  libxsmm_matrix_eqn_push_back_arg( my_eqn11, res.bc, res.H*res.W/res.num_HW_blocks, ld, 0, 0, in_dt );          /* inp [HW, bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn11, res.bc, 1, 1, 1, 0, LIBXSMM_DATATYPE_F32 );           /* a [bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn11, res.bc, 1, 1, 2, 0, LIBXSMM_DATATYPE_F32 );           /* b [bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn11, res.bc, res.H*res.W/res.num_HW_blocks, ld, 3, 0, in_dt );          /* dout [HW, bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn11, res.bc, 1, 1, 4, 0, LIBXSMM_DATATYPE_F32 );           /* dgamma [bc] */
-  /* libxsmm_matrix_eqn_tree_print( my_eqn11 ); */
-  /* libxsmm_matrix_eqn_rpn_print( my_eqn11 ); */
-  res.dgamma_func = libxsmm_dispatch_matrix_eqn( res.bc, 1, &tmp_ld2, LIBXSMM_DATATYPE_F32, my_eqn11 );      /* dgamma [bc] */
-  if ( res.dgamma_func == NULL) {
-    fprintf( stderr, "JIT for TPP bwd dgamma_func (eqn11) failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  /* dbeta function  */
-  my_eqn12 = libxsmm_matrix_eqn_create();                                                       /* dbeta [bc] = dout [HW, bc] + dbeta [bc] */
-  libxsmm_matrix_eqn_push_back_binary_op( my_eqn12, LIBXSMM_MELTW_TYPE_BINARY_ADD, LIBXSMM_MELTW_FLAG_BINARY_NONE, LIBXSMM_DATATYPE_F32 );                /* dbeta_tmp [HW, bc] */
-  libxsmm_matrix_eqn_push_back_unary_op(my_eqn12, LIBXSMM_MELTW_TYPE_UNARY_REDUCE_X_OP_ADD, LIBXSMM_MELTW_FLAG_UNARY_REDUCE_COLS, LIBXSMM_DATATYPE_F32);  /* [HW, bc] -> [bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn12, res.bc, res.H*res.W/res.num_HW_blocks, ld, 3, 0, in_dt );          /* dout [HW, bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn12, res.bc, 1, 1, 5, 0, LIBXSMM_DATATYPE_F32 );           /* dbeta [bc] */
-  /* libxsmm_matrix_eqn_tree_print( my_eqn12 ); */
-  /* libxsmm_matrix_eqn_rpn_print( my_eqn12 ); */
-  res.dbeta_func = libxsmm_dispatch_matrix_eqn( res.bc, 1, &tmp_ld2, LIBXSMM_DATATYPE_F32, my_eqn12 );      /* dbeta [bc] */
-  if ( res.dbeta_func == NULL) {
-    fprintf( stderr, "JIT for TPP bwd dbeta_func (eqn12) failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  /* din = gamma_ptr[v] * brstd_ptr[v] * recp_nhw * (nhw*del_output_ptr[v] - (del_beta_ptr[v] + (input_ptr[v] - bmean_ptr[v]) * del_gamma_ptr[v] * brstd_ptr[v])) */
-  /* din = gamma_ptr[v] * brstd_ptr[v] *del_output_ptr[v] - gamma_ptr[v] * brstd_ptr[v] * recp_nhw * (del_beta_ptr[v] + (input_ptr[v] - bmean_ptr[v]) * del_gamma_ptr[v] * brstd_ptr[v])) */
-  /* din = gamma_ptr[v] * brstd_ptr[v] *del_output_ptr[v] - gamma_ptr[v] * brstd_ptr[v] * recp_nhw * del_beta_ptr[v] + gamma_ptr[v] * brstd_ptr[v] * recp_nhw * (input_ptr[v] - bmean_ptr[v]) * del_gamma_ptr[v] * brstd_ptr[v]) */
-  /* din = a * del_output_ptr[v] + b * input_ptr[v] + c */
-  /* a = gamma_ptr[bc] * brstd_ptr[bc] */
-  /* b = gamma_ptr[bc] *  del_gamma_ptr[v] * brstd_ptr[bc] * brstd_ptr[bc] * recp_nhw */
-  /* c = -gamma_ptr[bc] * brstd_ptr[bc] * recp_nhw * del_beta_ptr[bc] + gamma_ptr[bc] * brstd_ptr[bc] * recp_nhw * bmean_ptr[bc] * del_gamma_ptr[bc] * brstd_ptr[bc]) */
-
-  /* din long equation */
-  my_eqn16 = libxsmm_matrix_eqn_create();                                                       /* din = a * dout + (b * inp + c) */
-  libxsmm_matrix_eqn_push_back_ternary_op( my_eqn16, LIBXSMM_MELTW_TYPE_TERNARY_MULADD, LIBXSMM_MELTW_FLAG_TERNARY_BCAST_COL_IN_0 | LIBXSMM_MELTW_FLAG_TERNARY_REUSE_IN_2_AS_OUT, LIBXSMM_DATATYPE_F32 );
-  libxsmm_matrix_eqn_push_back_arg( my_eqn16, res.bc, 1, 1, 1, 0, LIBXSMM_DATATYPE_F32 );           /* a [bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn16, res.bc, res.H*res.W/res.num_HW_blocks, ld, 3, 0, in_dt );          /* dout [HW, bc] */
-  libxsmm_matrix_eqn_push_back_ternary_op( my_eqn16, LIBXSMM_MELTW_TYPE_TERNARY_MULADD, LIBXSMM_MELTW_FLAG_TERNARY_BCAST_COL_IN_1 | LIBXSMM_MELTW_FLAG_TERNARY_BCAST_COL_IN_2 | LIBXSMM_MELTW_FLAG_TERNARY_REUSE_IN_2_AS_OUT, LIBXSMM_DATATYPE_F32 );
-  libxsmm_matrix_eqn_push_back_arg( my_eqn16, res.bc, res.H*res.W/res.num_HW_blocks, ld, 0, 0, in_dt );          /* inp [HW, bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn16, res.bc, 1, 1, 2, 0, LIBXSMM_DATATYPE_F32 );           /* b [bc] */
-  libxsmm_matrix_eqn_push_back_arg( my_eqn16, res.bc, 1, 1, 7, 0, LIBXSMM_DATATYPE_F32 );           /* c [bc] */
-  /* libxsmm_matrix_eqn_tree_print( my_eqn16 ); */
-  /* libxsmm_matrix_eqn_rpn_print( my_eqn16 ); */
-  res.din_func = libxsmm_dispatch_matrix_eqn( res.bc, res.H*res.W/res.num_HW_blocks, &ld, in_dt, my_eqn16 );           /* din [HW, bc] */
-  if ( res.din_func == NULL) {
-    fprintf( stderr, "JIT for TPP bwd din_func (eqn16) failed. Bailing...!\n");
-    exit(-1);
-  }
-
-  /* init scratch */
-  dbeta_N_offset = LIBXSMM_UP2(res.CP * res.N * res.bc, 64);
-  res.scratch_size =  sizeof(float) * ( dbeta_N_offset /* dbeta_N*/ + LIBXSMM_UP2(res.CP * res.N * res.bc, 64) /*dgamma_N */ );
-
-  return res;
-}
-
-void my_bn_bwd_exec( my_bn_bwd_config cfg, float *pdout, const float *pinp, const float *mean, const float *var, const float *pgamma, const unsigned char *prelumask,
-                     float *pdin, float *pdin_add, float *pdgamma, float *pdbeta, float eps,
-                     int start_tid, int my_tid, void *scratch) {
-
-  const libxsmm_blasint N  = cfg.N;
-  const libxsmm_blasint CP = cfg.CP;
-  const libxsmm_blasint HW = cfg.H * cfg.W;
-  const libxsmm_blasint bc = cfg.bc;
-  const libxsmm_blasint num_HW_blocks = cfg.num_HW_blocks;
-
-  /* computing first logical thread */
-  const libxsmm_blasint ltid = my_tid - start_tid;
-
-  /* number of tasks that could be run in parallel for 1d blocking */
-  /* Question: each thread should take a number of full (of length CP chunks) or can we really do a partial split here? */
-  const libxsmm_blasint work_dN = N * CP;
-  /* compute chunk size */
-  const libxsmm_blasint chunksize_dN = (work_dN % cfg.threads == 0) ?
-    (work_dN / cfg.threads) : ((work_dN / cfg.threads) + 1);
-  /* compute thr_begin and thr_end */
-  const libxsmm_blasint thr_begin_dN = ( ltid      * chunksize_dN < work_dN) ? ( ltid      * chunksize_dN) : work_dN;
-  const libxsmm_blasint thr_end_dN   = ((ltid + 1) * chunksize_dN < work_dN) ? ((ltid + 1) * chunksize_dN) : work_dN;
-
-  /* number of tasks that could be run in parallel for 1d blocking */
-  /* Question: each thread should take a number of full (of length CP chunks) or can we really do a partial split here? */
-  const libxsmm_blasint work_C = CP;
-  /* compute chunk size */
-  const libxsmm_blasint chunksize_C = (work_C % cfg.threads == 0) ?
-    (work_C / cfg.threads) : ((work_C / cfg.threads) + 1);
-  /* compute thr_begin and thr_end */
-  const libxsmm_blasint thr_begin_C = ( ltid      * chunksize_C < work_C) ? ( ltid      * chunksize_C) : work_C;
-  const libxsmm_blasint thr_end_C   = ((ltid + 1) * chunksize_C < work_C) ? ((ltid + 1) * chunksize_C) : work_C;
-
-  /* lazy barrier init */
-  libxsmm_barrier_init(cfg.barrier, ltid);
-
-  const float scale = 1.0f / ((float)N*HW);                   /* Scaling parameter*/
-
-  LIBXSMM_VLA_DECL(4,       float, din, pdin, CP, HW, bc);          /* [N, CP, HW, bc] */
-  LIBXSMM_VLA_DECL(4, const float, inp, pinp, CP, HW, bc);          /* [N, CP, HW, bc] */
-  LIBXSMM_VLA_DECL(4,       float, dout, pdout, CP, HW, bc);        /* [N, CP, HW, bc] */
-  LIBXSMM_VLA_DECL(2, const float, gamma, pgamma, bc);              /* [CP, bc] */
-  LIBXSMM_VLA_DECL(2,       float, dgamma, pdgamma, bc);            /* [CP, bc] */
-  LIBXSMM_VLA_DECL(2,       float, dbeta, pdbeta, bc);              /* [CP, bc] */
-
-  LIBXSMM_VLA_DECL(4,       float, din_add, pdin_add, CP, HW, bc);          /* [N, CP, HW, bc] */
-
-  float alpha = 0.0f;
-  LIBXSMM_VLA_DECL(4,       unsigned char, relumask, prelumask, CP, HW, bc/BITS_PER_CHAR);    /* [N, CP, HW, bc/BITS_PER_CHAR] */
-
-  const libxsmm_blasint dbeta_N_offset = (LIBXSMM_UP2((uintptr_t)(((float*)scratch) + CP * N * bc), 64) - ((uintptr_t)(scratch))) / sizeof(float);
-  LIBXSMM_VLA_DECL(3, float, dgamma_N, ((float*)scratch),                  N, bc);  /* [CP, N, bc] */
-  LIBXSMM_ASSUME_ALIGNED(dgamma_N_, 64);
-  LIBXSMM_VLA_DECL(3, float, dbeta_N,  ((float*)scratch) + dbeta_N_offset, N, bc);  /* [CP, N, bc] */
-  LIBXSMM_ASSUME_ALIGNED(dbeta_N_, 64);
-
-  { /* stupid block to keep indentation */
-    LIBXSMM_ALIGNED(float a[bc], 64); /* could also get moved into the scratch but left on the private stack as these are small, same below */
-    LIBXSMM_ALIGNED(float b[bc], 64);
-    LIBXSMM_ALIGNED(float c[bc], 64);
-    int n, cp;
-
-    int cpxnt;
-
-    /* ReLU/Mask/Eltwise */
-    if (cfg.fuse_type == MY_NORMALIZE_FUSE_ELTWISE ||
-        cfg.fuse_type == MY_NORMALIZE_FUSE_RELU || cfg.fuse_type == MY_NORMALIZE_FUSE_RELU_WITH_MASK || cfg.fuse_type == MY_NORMALIZE_FUSE_ELTWISE_RELU_WITH_MASK) {
-
-      for ( cpxnt = thr_begin_dN; cpxnt < thr_end_dN; ++cpxnt ) {
-        { /* stupid block to keep indentation */
-          n  = cpxnt%N;
-          cp = cpxnt/N;
-
-          int hwb, cb;
-
-          for(hwb=0; hwb < num_HW_blocks; hwb++){
-            if (cfg.fuse_type == MY_NORMALIZE_FUSE_RELU || cfg.fuse_type == MY_NORMALIZE_FUSE_RELU_WITH_MASK || cfg.fuse_type == MY_NORMALIZE_FUSE_ELTWISE_RELU_WITH_MASK) {
-              libxsmm_meltw_unary_param all_relu_param;
-
-              all_relu_param.op.primary   = (void*)(&alpha);
-              all_relu_param.in.primary   = &LIBXSMM_VLA_ACCESS(4, dout, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);      /* [HW,bc] */
-              all_relu_param.in.secondary = ((cfg.fuse_type == MY_NORMALIZE_FUSE_RELU_WITH_MASK || cfg.fuse_type == MY_NORMALIZE_FUSE_ELTWISE_RELU_WITH_MASK) ?
-                                               (void*)&LIBXSMM_VLA_ACCESS(4, relumask, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc/8)
-                                               : NULL /*&LIBXSMM_VLA_ACCESS(4, dout, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc) */ ); /* dout_fwd ? nonsense? */
-              all_relu_param.out.primary  = &LIBXSMM_VLA_ACCESS(4, dout, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);      /* [HW,bc] */
-              cfg.relu_kernel(&all_relu_param);
-            } /* ReLU/mask */
-//#if 0
-            if (cfg.fuse_type == MY_NORMALIZE_FUSE_ELTWISE || cfg.fuse_type == MY_NORMALIZE_FUSE_ELTWISE_RELU_WITH_MASK) {
-#if 0
-              int i;
-              for (i = 0; i < bc * (HW/num_HW_blocks); ++i) {
-                int index;
-                index = n * (CP * HW * bc) + cp * (HW * bc) + (hwb*(HW/num_HW_blocks)) * (bc) + i;
-                pdin_add[index] = pdout[index];
-              }
-#endif
-//#if 0
-              libxsmm_meltw_unary_param ewise_copy_param;
-              ewise_copy_param.in.primary  = &LIBXSMM_VLA_ACCESS(4, dout,    n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);
-              ewise_copy_param.out.primary = &LIBXSMM_VLA_ACCESS(4, din_add, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);
-              cfg.ewise_copy_kernel(&ewise_copy_param);
-//#endif
-            } /* Eltwise */
-//#endif
-          }
-        }
-      } /* loop over the 1d parallel blocking */
-    }  /* ReLU/Mask/Eltwise */
-
-    libxsmm_barrier_wait(cfg.barrier, ltid);
-
-    for ( cpxnt = thr_begin_dN; cpxnt < thr_end_dN; ++cpxnt ) {
-      { /* stupid block to keep indentation */
-        n  = cpxnt%N;
-        cp = cpxnt/N;
-
-        int hwb, cb;
-        libxsmm_matrix_arg arg_array[10];                                                           /* Private values of args and params */
-        libxsmm_matrix_eqn_param eqn_param;
-
-        LIBXSMM_ALIGNED(float lcl_dgamma_ptr[bc], 64);
-        LIBXSMM_ALIGNED(float lcl_dbeta_ptr[bc], 64);
-
-        float *dgamma_ncp_ptr = &LIBXSMM_VLA_ACCESS(3, dgamma_N, cp, n, 0, N, bc);
-        float *dbeta_ncp_ptr  = &LIBXSMM_VLA_ACCESS(3, dbeta_N, cp, n, 0, N, bc);
-
-        libxsmm_meltw_unary_param all_zero_param;
-        all_zero_param.out.primary = lcl_dgamma_ptr;
-        cfg.all_zero_kernel(&all_zero_param);
-        all_zero_param.out.primary = lcl_dbeta_ptr;
-        cfg.all_zero_kernel(&all_zero_param);
-
-        /* #pragma omp simd */
-        /* for (int cb = 0; cb < bc; cb++) { */
-        /*   lcl_dgamma_ptr[cb] = 0.0f; */
-        /*   lcl_dbeta_ptr[cb] = 0.0f; */
-        /* } */
-
-        for(cb = 0; cb < bc; cb++){
-          a[cb] = 1.0f / ((float)sqrt(var[cp*bc + cb] + eps));
-          b[cb] = -a[cb]*mean[cp*bc + cb];
-        }
-
-        arg_array[1].primary = a;
-        arg_array[2].primary = b;
-        arg_array[4].primary = lcl_dgamma_ptr;
-        arg_array[5].primary = lcl_dbeta_ptr;
-        arg_array[6].primary = (void*)&LIBXSMM_VLA_ACCESS(2, gamma, cp, 0, bc);
-
-        for(hwb=0; hwb < num_HW_blocks; hwb++){
-
-          arg_array[0].primary = (void*)&LIBXSMM_VLA_ACCESS(4, inp, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);
-          arg_array[3].primary = (void*)&LIBXSMM_VLA_ACCESS(4, dout, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);
-
-          eqn_param.inputs = arg_array;
-          eqn_param.output.primary = lcl_dgamma_ptr;
-          cfg.dgamma_func(&eqn_param);                                                             /* dgamma += (a * inp + b) * dout */
-
-          eqn_param.output.primary = lcl_dbeta_ptr;
-          cfg.dbeta_func(&eqn_param);                                                              /* dbeta += dout */
-        }
-
-        libxsmm_meltw_unary_param copy_param;
-        copy_param.in.primary = lcl_dgamma_ptr;
-        copy_param.out.primary = dgamma_ncp_ptr;
-        cfg.copy_kernel(&copy_param);
-
-        copy_param.in.primary = lcl_dbeta_ptr;
-        copy_param.out.primary = dbeta_ncp_ptr;
-        cfg.copy_kernel(&copy_param);
-
-        /* #pragma omp simd */
-        /* for (int cb = 0; cb < bc; cb++) { */
-        /*   dgamma_ncp_ptr[cb] = lcl_dgamma_ptr[cb]; */
-        /*   dbeta_ncp_ptr[cb] = lcl_dbeta_ptr[cb]; */
-        /* } */
-      }
-    }
-
-    libxsmm_barrier_wait(cfg.barrier, ltid);
-
-    for ( cp = thr_begin_C; cp < thr_end_C; ++cp ) {
-      libxsmm_meltw_unary_param all_zero_param;
-      all_zero_param.out.primary = &LIBXSMM_VLA_ACCESS(2, dgamma, cp, 0, bc);
-      cfg.all_zero_kernel(&all_zero_param);
-      all_zero_param.out.primary = &LIBXSMM_VLA_ACCESS(2, dbeta, cp, 0, bc);
-      cfg.all_zero_kernel(&all_zero_param);
-
-      /* #pragma omp simd */
-      /* for (int cb = 0; cb < bc; cb++) { */
-      /*   pdgamma[cp*bc + cb] = 0.0f; */
-      /*   pdbeta[cp*bc + cb] = 0.0f; */
-      /* } */
-
-      libxsmm_meltw_binary_param add_param;
-      int ni;
-      for(ni = 0; ni < N; ni++){
-
-        add_param.in0.primary = &LIBXSMM_VLA_ACCESS(2, dgamma, cp, 0, bc);
-        add_param.in1.primary = &LIBXSMM_VLA_ACCESS(3, dgamma_N, cp, ni, 0, N, bc);
-        add_param.out.primary = &LIBXSMM_VLA_ACCESS(2, dgamma, cp, 0, bc);
-        cfg.add_kernel(&add_param);
-
-        add_param.in0.primary = &LIBXSMM_VLA_ACCESS(2, dbeta, cp, 0, bc);
-        add_param.in1.primary = &LIBXSMM_VLA_ACCESS(3, dbeta_N, cp, ni, 0, N, bc);
-        add_param.out.primary = &LIBXSMM_VLA_ACCESS(2, dbeta, cp, 0, bc);
-        cfg.add_kernel(&add_param);
-
-        /* #pragma omp simd */
-        /* for (int cb = 0; cb < bc; cb++) { */
-        /*   pdgamma[cp*bc + cb] += dgamma_N[cp*N*bc + n*bc + cb];  */
-        /*   pdbeta[cp*bc + cb] += dbeta_N[cp*N*bc + n*bc + cb];  */
-        /* } */
-      }
-    }
-
-    libxsmm_barrier_wait(cfg.barrier, ltid);
-
-    for ( cpxnt = thr_begin_dN; cpxnt < thr_end_dN; ++cpxnt ) {
-      { /* stupid block to keep indentation */
-        n  = cpxnt%N;
-        cp = cpxnt/N;
-
-        libxsmm_matrix_arg arg_array[8];                                                               /* Private eqn args and params */
-        libxsmm_matrix_eqn_param eqn_param;
-        int hwb, cb;
-
-        /* FIXME: Replace expressions for pgamma, pdgamma etc. with ACCESS? */
-        for(cb = 0; cb < bc; cb++){
-          a[cb] = pgamma[cp*bc + cb] / ((float)sqrt(var[cp*bc + cb] + eps));                            /* a = gamma_ptr[bc] * brstd_ptr[bc] */
-          b[cb] = -a[cb] * scale * pdgamma[cp*bc + cb] / ((float)sqrt(var[cp*bc + cb] + eps));          /* b = gamma_ptr[bc] * brstd_ptr[bc] * del_gamma_ptr[v] * brstd_ptr[bc] * recp_nhw */
-          c[cb] = -b[cb] * mean[cp*bc + cb] - a[cb] * scale * pdbeta[cp*bc + cb] ;                      /* c = -gamma_ptr[bc] * brstd_ptr[bc] * recp_nhw * del_beta_ptr[bc] + gamma_ptr[bc] * brstd_ptr[bc] * recp_nhw * bmean_ptr[bc] * del_gamma_ptr[bc] * brstd_ptr[bc]) */
-        }
-
-        arg_array[1].primary = a;
-        arg_array[2].primary = b;
-        arg_array[6].primary = (void*)&LIBXSMM_VLA_ACCESS(2, gamma, cp, 0, bc);
-        arg_array[7].primary = c;
-
-        for(hwb=0; hwb < num_HW_blocks; hwb++){
-          arg_array[0].primary = (void*)&LIBXSMM_VLA_ACCESS(4, inp, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);
-          arg_array[3].primary = (void*)&LIBXSMM_VLA_ACCESS(4, dout, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);
-
-          eqn_param.inputs = arg_array;
-          eqn_param.output.primary = &LIBXSMM_VLA_ACCESS(4, din, n, cp, hwb*(HW/num_HW_blocks), 0, CP, HW, bc);
-          cfg.din_func(&eqn_param);                                                                        /* din = dout * a + b * inp + c */
-
-        }
-      }
-    }
-  } /* simple code block or parallel section for old code */
-
-  libxsmm_barrier_wait(cfg.barrier, ltid);
-}
-
-
-#ifndef CNN_HEADER
 
 int main( int argc, char* argv[] ) {
 
@@ -1481,6 +687,4 @@ int main( int argc, char* argv[] ) {
 
   return 0;
 }
-
-#endif /* for #ifndef CNN_HEADER */
