+++ version.txt
@@ -1,1 +1,1 @@
-dev/kvoronin/batchnormalization_tpp-1.17-2208
+dev/kvoronin/batchnormalization_tpp-1.17-2209
